requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tk==0.1.0
customtkinter==5.2.1
